feat: 페이징처리,유저 접근 수정, id반환 수정, 특정게시물 로직 수정,update, delete처리 구현

+++ backend/src/main/java/hello/aimju/Comment/service/CommentService.java
@@ -2,8 +2,12 @@ package hello.aimju.Comment.service;
 
 
 import hello.aimju.Board.domain.Board;
+import hello.aimju.Board.dto.ModifyBoardRequestDto;
+import hello.aimju.Board.dto.ModifyBoardResponseDto;
 import hello.aimju.Board.repository.BoardRepository;
 import hello.aimju.Comment.domain.Comment;
+import hello.aimju.Comment.dto.ModifyCommentRequestDto;
+import hello.aimju.Comment.dto.ModifyCommentResponseDto;
 import hello.aimju.Comment.dto.RetrieveCommentsResponseDto;
 import hello.aimju.Comment.dto.WriteCommentRequestDto;
 import hello.aimju.Comment.repository.CommentRepository;
@@ -42,13 +46,33 @@ public class CommentService {
         List<Comment> boardComments = commentRepository.findAllByBoard(board);
         List<RetrieveCommentsResponseDto> getCommentLists = boardComments.stream()
                 .map(tmp -> RetrieveCommentsResponseDto.builder()
+                        .id(tmp.getId())
                         .name(tmp.getUser().getUserName())
                         .content(tmp.getContent())
                         .build())
                 .collect(Collectors.toList());
         return new ResponseEntity<>(getCommentLists,HttpStatus.OK);
     }
-    public void deleteComment(Long commentId) {
+    public ResponseEntity<?> modifyBoard(ModifyCommentRequestDto modifyCommentRequestDto, HttpSession session) throws Exception {
+        Comment comment = commentRepository.findById(modifyCommentRequestDto.getId()).get();
+        checkUser(getUserFromSession(session), comment.getUser());
+        comment.update(modifyCommentRequestDto.getContent());
+        ModifyCommentResponseDto resComment = ModifyCommentResponseDto.builder()
+                .id(comment.getId())
+                .content(comment.getContent())
+                .build();
+        commentRepository.save(comment);
+        return new ResponseEntity<>(resComment, HttpStatus.OK);
+    }
+    public void deleteComment(Long commentId, HttpSession session) throws Exception {
+        User userFromSession = getUserFromSession(session);
+        Comment comment = commentRepository.findById(commentId).get();
+        User user = comment.getUser();
+        System.out.println("userFromSession.getId() = " + userFromSession.getId());
+        System.out.println("user.getId() = " + user.getId());
+        if (userFromSession.getId() != user.getId()){
+            throw new Exception("자신의 댓글만 삭제할 수 있어요");
+        }
         commentRepository.deleteById(commentId);
     }
 
@@ -62,5 +86,10 @@ public class CommentService {
         }
 
         return loginUser;
+    }
+    private void checkUser(User live, User save) throws Exception {
+        if (live.getId() != save.getId()){
+            throw new Exception("자신의 댓글만 삭제할 수 있어요");
+        }
     }
 }
